route/cls: U32 classifier bugfixes
1) Fix u32_clone() so it makes proper cloning of action.
Do not do raw memcpy. It was causing double freeing of memory and
reference count errors.
For now, only the first action in the action chain is included
in the clone. This may need to be fixed also.

2) Fix rtnl_u32_get_action() so it no longer does a raw memcpy
of the action object. Instead the reference counter is increased
to keep the original object alive. The action returned by this
function must be "put" after use.

Signed-off-by: Magnus Öberg <magnus.oberg@westermo.se>