route/link: add support for bridge specific settings
The bridge support in libnl has since the start focused mainly (only?)
on the bridge links, or ports, not the bridge itself.  This introduces
complete rtnl_link_info_ops for the bridge and implements two new APIs
to create a VLAN filterin bridge and set which PVID new links enslaved
to the bridge will get by default.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>