route/link: add VLAN user carrier flag
This is a Westermo extension to the kernel carrier handling for VLAN
interfaces.  The intention is to bypass any automatic carrier handling
particularly for when an interface should reamin "always up" and not
deconfigure any acquired DCHP lease, for instance.  On unstable links,
or when networks reconfigure, we don't want temporary (less than 0.5s)
link loss on bridge ports to cause VLAN interfas to go down.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>