[WMO]lib: compat nl_cache_alloc() with libnl v3.5.0 behavior
We use the cache concept in libnl to a) get a bunch of links from the
kernel of currently active interfaces, and b) create all links we want
after a WeOS reconfiguration (e.g. new config file).  We then ask libnl
to compare the caches, using our __diff() callbacks, so we can figure
out exactly which interfaces need to be changed and/or added -- this
allows us to minimize changes the end-user experiences.  Which in many
cases is very important; not having link changes to interfaces you have
not changed may seem obvious -- the cache handling is what provides that
in WeOsS.

However, with the changes in libnl v3.6.0 we can no loger add our new
interfaces to a cache because they all had the same (unset) ifindex, and
instead of making an exception for new interfaces we decided to add this
compat cache behavior.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>