link: Add function to unset carrier request
When fetching from the kernel this flag seems to be set and when
making a change request this persists. Since it's only internal
variables we need a function to unset it from the outside.

ndo_carrier_change in the kernel should only be implemented by
non-hardware interfaces. For hardware interfaces (e.g. physical ports)
it will be EOPNOTSUPP. Allow unsetting it to not get an error.

Signed-off-by: Casper Andersson <casper.andersson@westermo.com>